# Conflicts:
#	.travis.yml
#	documentation/changelog.creole
#	pom.xml
#	src/main/java/org/mariadb/jdbc/MariaDbConnection.java
#	src/main/java/org/mariadb/jdbc/MariaDbStatement.java
#	src/main/java/org/mariadb/jdbc/internal/com/read/resultset/rowprotocol/BinaryRowProtocol.java
#	src/main/java/org/mariadb/jdbc/internal/failover/FailoverProxy.java
#	src/main/java/org/mariadb/jdbc/internal/util/constant/Version.java
#	src/main/java/org/mariadb/jdbc/internal/util/pool/Pool.java
#	src/test/java/org/mariadb/jdbc/DriverTest.java
#	src/test/java/org/mariadb/jdbc/GeometryTest.java
#	src/test/java/org/mariadb/jdbc/ReconnectionStateMaxAllowedStatement.java
#	src/test/java/org/mariadb/jdbc/TimeoutTest.java
#	src/test/java/org/mariadb/jdbc/failover/AuroraAutoDiscoveryTest.java
#	src/test/java/org/mariadb/jdbc/failover/AuroraFailoverTest.java
#	src/test/java/org/mariadb/jdbc/failover/GaleraFailoverTest.java

+++ src/test/java/org/mariadb/jdbc/TransactionTest.java
@@ -52,14 +52,14 @@
 
 package org.mariadb.jdbc;
 
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+
 import java.sql.Connection;
-import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-import org.junit.*;
-
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
@@ -73,14 +73,16 @@ public class TransactionTest extends BaseTest {
      */
     @Before
     public void before() throws SQLException {
-        Statement stmt = sharedConnection.createStatement();
-        stmt.execute("drop table if exists tx_fore_key");
-        stmt.execute("drop table if exists tx_prim_key");
-        createTable("tx_prim_key", "id int not null primary key", "engine=innodb");
-        createTable("tx_fore_key",
-                "id int not null primary key, id_ref int not null, "
-                        + "foreign key (id_ref) references tx_prim_key(id) on delete restrict on update restrict",
-                "engine=innodb");
+        if (testSingleHost) {
+            Statement stmt = sharedConnection.createStatement();
+            stmt.execute("drop table if exists tx_fore_key");
+            stmt.execute("drop table if exists tx_prim_key");
+            createTable("tx_prim_key", "id int not null primary key", "engine=innodb");
+            createTable("tx_fore_key",
+                    "id int not null primary key, id_ref int not null, "
+                            + "foreign key (id_ref) references tx_prim_key(id) on delete restrict on update restrict",
+                    "engine=innodb");
+        }
     }
 
     /**
@@ -91,9 +93,11 @@ public class TransactionTest extends BaseTest {
      */
     @After
     public void after() throws SQLException {
-        Statement stmt = sharedConnection.createStatement();
-        stmt.execute("drop table if exists tx_fore_key");
-        stmt.execute("drop table if exists tx_prim_key");
+        if (testSingleHost) {
+            Statement stmt = sharedConnection.createStatement();
+            stmt.execute("drop table if exists tx_fore_key");
+            stmt.execute("drop table if exists tx_prim_key");
+        }
     }
 
     @Test
